Server/requirements.txt
async-timeout==4.0.2
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
Faker==17.5.0
Flask==2.2.3
Flask-SocketIO==5.3.2
fonttools==4.38.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
numpy==1.21.6
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
redis==4.5.1
requests==2.28.2
scipy==1.7.3
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.15.0
